- initial setup of the FilterFactory
